refactor(analysis): split architecture helpers
