+++ skylos/analysis/architecture.py
@@ -10,11 +10,12 @@ Rule IDs:
 from __future__ import annotations
 
 import ast
-import fnmatch
 from collections import defaultdict
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
+from skylos.analysis.architecture_findings import generate_architecture_findings
+from skylos.analysis.architecture_layers import get_layer_policy_findings
 
 
 @dataclass
@@ -408,418 +409,12 @@ def analyze_architecture(
             "zone_distribution": {},
         }
 
-    _generate_findings(result, iad_findings_advisory=iad_findings_advisory)
-
-    return result
-
-
-def _iad_scope_fields(advisory: bool) -> dict[str, Any]:
-    fields = {
-        "advisory": advisory,
-        "scope": "file",
-        "metric_granularity": "file-level heuristic",
-        "metric_origin": "Martin I/A/D release-unit metric",
-    }
-    if advisory:
-        fields["advisory_reason"] = (
-            "Martin's instability/abstractness/distance metric was defined for "
-            "release units. Skylos reports it at Python file granularity as an "
-            "architecture signal; it is not gate-blocking unless I/A/D enforcement "
-            "is enabled."
-        )
-    else:
-        fields["enforcement_reason"] = (
-            "I/A/D enforcement is enabled, so this file-level architecture signal "
-            "can block strict gates."
-        )
-    return fields
-
-
-def _iad_gate_note(advisory: bool) -> str:
-    if advisory:
-        return (
-            "Advisory: this file-level I/A/D signal does not block gates unless "
-            "I/A/D enforcement is enabled. See remediations for safe options."
-        )
-    return (
-        "I/A/D enforcement is enabled, so this file-level architecture signal can "
-        "block strict gates. See remediations for safe options."
+    generate_architecture_findings(
+        result,
+        iad_findings_advisory=iad_findings_advisory,
     )
 
-
-def _iad_docs_url(rule_id: str) -> str:
-    return f"https://docs.skylos.dev/rules/{rule_id}"
-
-
-def _iad_remediations(module_name: str, zone: str) -> list[dict[str, str]]:
-    simple_name = module_name.rsplit(".", 1)[-1]
-    if simple_name.startswith("_"):
-        helper_hint = (
-            "This module is already marked private. If it belongs to one release-unit, "
-            "keep the finding as an advisory signal unless the fan-in reflects a real "
-            "maintenance problem. Do not add fake imports or one-off abstractions just "
-            "to move the metric."
-        )
-        helper_title = "Keep private-helper intent explicit"
-    else:
-        private_name = f"_{simple_name}"
-        helper_hint = (
-            f"If '{module_name}' is an internal helper rather than public API, rename "
-            f"the module to '{private_name}' and update imports. Do not add fake "
-            "imports or one-off abstractions just to move the metric."
-        )
-        helper_title = "Mark internal helpers as private"
-
-    remediations = [
-        {
-            "title": helper_title,
-            "hint": helper_hint,
-            "applies_when": "The module is an implementation detail inside one release-unit.",
-        },
-        {
-            "title": "Introduce a real abstraction boundary",
-            "hint": (
-                "Add a Protocol or ABC only when consumers genuinely need an interface "
-                "with multiple implementations or an extension boundary. Avoid header "
-                "interfaces created only to satisfy the score."
-            ),
-            "applies_when": "The module is public API or has real interchangeable implementations.",
-        },
-        {
-            "title": "Split responsibilities when fan-in is meaningful",
-            "hint": (
-                "If many consumers depend on unrelated behavior in this module, split "
-                "cohesive responsibilities so changes do not ripple through all callers."
-            ),
-            "applies_when": "High fan-in reflects mixed responsibilities, not just shared helpers.",
-        },
-    ]
-
-    if zone == "zone_of_uselessness":
-        remediations.append(
-            {
-                "title": "Remove or move unused abstractions",
-                "hint": (
-                    "If abstractions have few stable consumers, move them near their "
-                    "callers or remove abstractions that are not carrying a real boundary."
-                ),
-                "applies_when": "The module is abstract but not depended on by stable code.",
-            }
-        )
-
-    return remediations
-
-
-def _generate_findings(
-    result: ArchitectureResult,
-    *,
-    iad_findings_advisory: bool = True,
-):
-    for name, m in result.modules.items():
-        # SKY-Q802: High distance from main sequence
-        if m.distance > 0.5 and m.total_coupling > 0:
-            if m.distance > 0.7:
-                severity = "HIGH"
-            else:
-                severity = "MEDIUM"
-
-            result.findings.append(
-                {
-                    "rule_id": "SKY-Q802",
-                    "kind": "architecture",
-                    "severity": severity,
-                    "type": "module",
-                    "name": name,
-                    "simple_name": name.split(".")[-1],
-                    "value": round(m.distance, 3),
-                    "threshold": 0.5,
-                    "instability": round(m.instability, 3),
-                    "abstractness": round(m.abstractness, 3),
-                    "message": (
-                        f"Module '{name}' is far from the Main Sequence "
-                        f"(D={m.distance:.2f}, I={m.instability:.2f}, A={m.abstractness:.2f}). "
-                        f"Zone: {m.zone.replace('_', ' ')}. "
-                        f"{_iad_gate_note(iad_findings_advisory)}"
-                    ),
-                    "file": m.file_path,
-                    "basename": Path(m.file_path).name if m.file_path else name,
-                    "line": 1,
-                    "docs_url": _iad_docs_url("SKY-Q802"),
-                    "remediations": _iad_remediations(name, m.zone),
-                    **_iad_scope_fields(iad_findings_advisory),
-                }
-            )
-
-        # SKY-Q803: Zone warnings
-        if m.zone in ("zone_of_pain", "zone_of_uselessness") and m.total_coupling > 0:
-            if m.zone == "zone_of_pain":
-                zone_msg = (
-                    f"Module '{name}' is in the Zone of Pain "
-                    f"(concrete A={m.abstractness:.2f}, stable I={m.instability:.2f}). "
-                    "Changes here can ripple widely at file granularity. "
-                    f"{_iad_gate_note(iad_findings_advisory)}"
-                )
-            else:
-                zone_msg = (
-                    f"Module '{name}' is in the Zone of Uselessness "
-                    f"(abstract A={m.abstractness:.2f}, unstable I={m.instability:.2f}). "
-                    "Few stable consumers depend on it. "
-                    f"{_iad_gate_note(iad_findings_advisory)}"
-                )
-
-            result.findings.append(
-                {
-                    "rule_id": "SKY-Q803",
-                    "kind": "architecture",
-                    "severity": "MEDIUM",
-                    "type": "module",
-                    "name": name,
-                    "simple_name": name.split(".")[-1],
-                    "value": m.zone,
-                    "instability": round(m.instability, 3),
-                    "abstractness": round(m.abstractness, 3),
-                    "distance": round(m.distance, 3),
-                    "message": zone_msg,
-                    "file": m.file_path,
-                    "basename": Path(m.file_path).name if m.file_path else name,
-                    "line": 1,
-                    "docs_url": _iad_docs_url("SKY-Q803"),
-                    "remediations": _iad_remediations(name, m.zone),
-                    **_iad_scope_fields(iad_findings_advisory),
-                }
-            )
-
-    # SKY-Q804: DIP violations
-    for v in result.dip_violations:
-        stable_file = result.modules.get(v.stable_module)
-        result.findings.append(
-            {
-                "rule_id": "SKY-Q804",
-                "kind": "architecture",
-                "severity": v.severity,
-                "type": "module",
-                "name": v.stable_module,
-                "simple_name": v.stable_module.split(".")[-1],
-                "value": f"{v.stable_module} -> {v.unstable_module}",
-                "message": (
-                    f"Dependency Inversion violation: stable module '{v.stable_module}' "
-                    f"(I={v.stable_instability:.2f}) depends on unstable module "
-                    f"'{v.unstable_module}' (I={v.unstable_instability:.2f}). "
-                    f"Consider introducing an abstraction layer."
-                ),
-                "file": stable_file.file_path if stable_file else "",
-                "basename": Path(stable_file.file_path).name
-                if stable_file and stable_file.file_path
-                else v.stable_module,
-                "line": 1,
-            }
-        )
-
-
-def _normalize_layer_policy(policy: Any) -> dict[str, Any] | None:
-    if not isinstance(policy, dict):
-        return None
-
-    raw_layers = policy.get("layers") or []
-    raw_rules = policy.get("rules") or []
-    if not isinstance(raw_layers, list) or not isinstance(raw_rules, list):
-        return None
-    if not raw_layers or not raw_rules:
-        return None
-
-    layers: list[dict[str, Any]] = []
-    for raw_layer in raw_layers:
-        if not isinstance(raw_layer, dict):
-            continue
-        name = str(raw_layer.get("name") or "").strip()
-        if not name:
-            continue
-        patterns = raw_layer.get("patterns")
-        if patterns is None:
-            patterns = raw_layer.get("packages")
-        if isinstance(patterns, str):
-            patterns = [patterns]
-        if not isinstance(patterns, list):
-            patterns = []
-        clean_patterns = [str(p).strip() for p in patterns if str(p).strip()]
-        if clean_patterns:
-            layers.append({"name": name, "patterns": clean_patterns})
-
-    rules: list[dict[str, Any]] = []
-    for raw_rule in raw_rules:
-        if not isinstance(raw_rule, dict):
-            continue
-        from_layer = str(raw_rule.get("from") or "").strip()
-        if not from_layer:
-            continue
-
-        allow = raw_rule.get("allow") or []
-        deny = raw_rule.get("deny") or []
-        if isinstance(allow, str):
-            allow = [allow]
-        if isinstance(deny, str):
-            deny = [deny]
-        if not isinstance(allow, list):
-            allow = []
-        if not isinstance(deny, list):
-            deny = []
-
-        clean_allow = [str(v).strip() for v in allow if str(v).strip()]
-        clean_deny = [str(v).strip() for v in deny if str(v).strip()]
-        if clean_allow or clean_deny:
-            rules.append(
-                {
-                    "from": from_layer,
-                    "allow": clean_allow,
-                    "deny": clean_deny,
-                    "severity": str(raw_rule.get("severity") or "HIGH").upper(),
-                }
-            )
-
-    if not layers or not rules:
-        return None
-
-    return {
-        "strict": bool(policy.get("strict", False)),
-        "layers": layers,
-        "rules": rules,
-    }
-
-
-def _layer_pattern_matches(module_name: str, file_path: str, pattern: str) -> bool:
-    normalized = pattern.strip().replace("/", ".").strip(".")
-    if not normalized:
-        return False
-
-    file_module = Path(file_path).with_suffix("").as_posix().replace("/", ".")
-    candidates = {module_name, file_module}
-    for candidate in candidates:
-        if candidate == normalized or candidate.startswith(normalized + "."):
-            return True
-        if fnmatch.fnmatchcase(candidate, normalized):
-            return True
-        if fnmatch.fnmatchcase(candidate, normalized + ".*"):
-            return True
-        if normalized in candidate.split("."):
-            return True
-
-    return False
-
-
-def _map_modules_to_layers(
-    module_files: dict[str, str],
-    layers: list[dict[str, Any]],
-) -> dict[str, str]:
-    mapped: dict[str, str] = {}
-    for module_name, file_path in module_files.items():
-        matches: list[tuple[int, int, str]] = []
-        for layer in layers:
-            for pattern in layer["patterns"]:
-                if _layer_pattern_matches(module_name, file_path, pattern):
-                    specificity = pattern.count(".")
-                    matches.append((specificity, len(pattern), layer["name"]))
-        if matches:
-            matches.sort(key=lambda item: (-item[0], -item[1], item[2]))
-            mapped[module_name] = matches[0][2]
-        else:
-            mapped[module_name] = "unknown"
-    return mapped
-
-
-def get_layer_policy_findings(
-    dependency_graph: dict[str, set[str]],
-    module_files: dict[str, str],
-    policy: Any,
-) -> tuple[list[dict[str, Any]], dict[str, Any]]:
-    normalized = _normalize_layer_policy(policy)
-    if not normalized:
-        return [], {}
-
-    layer_map = _map_modules_to_layers(module_files, normalized["layers"])
-    rules_by_from = {rule["from"]: rule for rule in normalized["rules"]}
-    findings: list[dict[str, Any]] = []
-
-    checked_edges = 0
-    for from_module, deps in sorted(dependency_graph.items()):
-        from_layer = layer_map.get(from_module, "unknown")
-        for to_module in sorted(deps):
-            if to_module not in module_files:
-                continue
-            to_layer = layer_map.get(to_module, "unknown")
-            if from_module == to_module:
-                continue
-            checked_edges += 1
-
-            rule = rules_by_from.get(from_layer)
-            violation_reason = ""
-            severity = "HIGH"
-
-            if normalized["strict"] and (
-                from_layer == "unknown" or to_layer == "unknown"
-            ):
-                violation_reason = (
-                    "Dependency involves module(s) outside configured architecture layers"
-                )
-                severity = "MEDIUM"
-            elif rule is None:
-                continue
-            elif to_layer in set(rule["deny"]):
-                violation_reason = (
-                    f"Layer '{from_layer}' is explicitly denied from depending on "
-                    f"'{to_layer}'"
-                )
-                severity = rule["severity"]
-            elif rule["allow"] and to_layer not in set(rule["allow"]):
-                allowed = ", ".join(rule["allow"])
-                violation_reason = (
-                    f"Layer '{from_layer}' may only depend on: {allowed}"
-                )
-                severity = rule["severity"]
-            else:
-                continue
-
-            source_file = module_files.get(from_module, "")
-            findings.append(
-                {
-                    "rule_id": "SKY-Q805",
-                    "kind": "architecture",
-                    "severity": severity
-                    if severity in {"CRITICAL", "HIGH", "MEDIUM", "LOW"}
-                    else "HIGH",
-                    "type": "module_dependency",
-                    "name": from_module,
-                    "simple_name": from_module.split(".")[-1],
-                    "value": f"{from_layer}->{to_layer}",
-                    "from_module": from_module,
-                    "to_module": to_module,
-                    "from_layer": from_layer,
-                    "to_layer": to_layer,
-                    "message": (
-                        f"Architecture layer violation: '{from_module}' "
-                        f"({from_layer}) depends on '{to_module}' ({to_layer}). "
-                        f"{violation_reason}."
-                    ),
-                    "file": source_file,
-                    "basename": Path(source_file).name if source_file else from_module,
-                    "line": 1,
-                }
-            )
-
-    mapped_counts: dict[str, int] = defaultdict(int)
-    for layer in layer_map.values():
-        mapped_counts[layer] += 1
-
-    summary = {
-        "enabled": True,
-        "strict": normalized["strict"],
-        "layer_count": len(normalized["layers"]),
-        "rule_count": len(normalized["rules"]),
-        "checked_edges": checked_edges,
-        "violation_count": len(findings),
-        "module_layers": dict(sorted(layer_map.items())),
-        "layer_distribution": dict(sorted(mapped_counts.items())),
-    }
-    return findings, summary
+    return result
 
 
 def get_architecture_findings(
